refactor to get things out of agenda.py
