+++ src/econ/test_econ_treasury/test_econ_treasury_voice_rank.py
@@ -1,8 +1,6 @@
-from src.agenda.agenda import (
-    agendaunit_shop,
-    get_file_names_in_voice_rank_order,
-    partyunit_shop,
-)
+from src.agenda.party import partyunit_shop
+from src.agenda.agenda import agendaunit_shop
+from src.agenda.meld_files import get_file_names_in_voice_rank_order
 from src._instrument.file import save_file, delete_dir
 from src.econ.econ import econunit_shop, save_role_file_agenda
 from src.econ.examples.econ_env_kit import env_dir_setup_cleanup, get_texas_econnox
